src/requirements.txt
bleak  # No version currently supports Python 3.13 as of September 2025. Remove or update when available.
bleak-winrt==1.2.0
contourpy==1.2.0
cycler==0.12.1
filterpy==1.4.5
fonttools==4.50.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==24.0
paho-mqtt==2.0.0
pillow
psycopg2-binary
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
# scipy==1.12.0
six==1.16.0
typing_extensions==4.11.0
